pyproject.toml
[project]
name = "temp-logger"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Milos Svana", email = "milos.svana@mailfence.com" }
]
dependencies = [
    "pyserial>=3.5",
    "setuptools>=68.1.2",
    "black>=23.7.0",
    "quart>=0.18.4",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
